requirements.txt
rdkit-pypi
pytorch
scikit-learn
numpy
pandas

